+++ src/EventSourcing/EventStores/EventStoreUserRepository.cs
@@ -2,16 +2,19 @@
 using ShopSharp.Core.Domain.Aggregates;
 using ShopSharp.Users.Domain.Aggregates;
 using ShopSharp.Users.Domain.Repositories;
+using ShopSharp.Users.EventSourcing.EventRecords;
 
-namespace ShopSharp.Users.EventSourcing;
+namespace ShopSharp.Users.EventSourcing.EventStores;
 
 internal class EventStoreUserRepository : IUserRepository
 {
+    private readonly IEventRecordFactory _eventRecordFactory;
     private readonly IEventStore _eventStore;
     private readonly IPublisher _publisher;
 
-    public EventStoreUserRepository(IEventStore eventStore, IPublisher publisher)
+    public EventStoreUserRepository(IEventRecordFactory eventRecordFactory, IEventStore eventStore, IPublisher publisher)
     {
+        _eventRecordFactory = eventRecordFactory;
         _eventStore = eventStore;
         _publisher = publisher;
     }
@@ -19,8 +22,9 @@ internal class EventStoreUserRepository : IUserRepository
     public async Task SaveAsync(User user, CancellationToken cancellationToken = default)
     {
         var streamId = GetStreamId(user);
+        var eventRecords = _eventRecordFactory.CreateFromDomainEvents(user.UncommittedDomainEvents);
 
-        await _eventStore.AppendToStreamAsync(streamId, user.UncommittedDomainEvents, cancellationToken)
+        await _eventStore.AppendToStreamAsync(streamId, eventRecords, cancellationToken)
             .ConfigureAwait(false);
 
         foreach (var domainEvent in user.UncommittedDomainEvents)